+++ src/Accounts/Accounts.Test/AutosaveTests.cs
@@ -13,10 +13,6 @@
 // ----------------------------------------------------------------------------------
 
 using Microsoft.Azure.Commands.Common.Authentication;
-// TODO: Remove IfDef
-#if NETSTANDARD
-using Microsoft.Azure.Commands.Common.Authentication.Core;
-#endif
 using Microsoft.Azure.Commands.Common.Authentication.Models;
 using Microsoft.Azure.ServiceManagement.Common.Models;
 using Microsoft.WindowsAzure.Commands.Common.Test.Mocks;
@@ -49,11 +45,15 @@ namespace Microsoft.Azure.Commands.Profile.Test
 
             TestExecutionHelpers.SetUpSessionAndProfile();
             ResourceManagerProfileProvider.InitializeResourceManagerProfile(true);
+            // prevent token acquisition
+            AzureRmProfileProvider.Instance.GetProfile<AzureRmProfile>().ShouldRefreshContextsFromCache = false;
             AzureSession.Instance.DataStore = dataStore;
             AzureSession.Instance.ARMContextSaveMode = ContextSaveMode.Process;
             AzureSession.Instance.AuthenticationFactory = new MockTokenAuthenticationFactory();
-            AzureSession.Instance.TokenCache = new AuthenticationStoreTokenCache(new AzureTokenCache());
             Environment.SetEnvironmentVariable("Azure_PS_Data_Collection", "false");
+            PowerShellTokenCacheProvider tokenProvider = new InMemoryTokenCacheProvider();
+            AzureSession.Instance.RegisterComponent(PowerShellTokenCacheProvider.PowerShellTokenCacheProviderKey, () => tokenProvider);
+            AzureSession.Instance.RegisterComponent(nameof(PowerShellTokenCache), () => tokenProvider.GetTokenCache());
         }
 
         [Fact]
@@ -70,8 +70,11 @@ namespace Microsoft.Azure.Commands.Profile.Test
                 cmdlet.ExecuteCmdlet();
                 cmdlet.InvokeEndProcessing();
                 Assert.Equal(ContextSaveMode.CurrentUser, AzureSession.Instance.ARMContextSaveMode);
-                Assert.Equal(typeof(ProtectedFileTokenCache), AzureSession.Instance.TokenCache.GetType());
                 Assert.Equal(typeof(ProtectedProfileProvider), AzureRmProfileProvider.Instance.GetType());
+            }
+            catch (PlatformNotSupportedException)
+            {
+                // swallow exception when test env (Linux server) doesn't support token cache persistence
             }
             finally
             {
@@ -94,8 +97,11 @@ namespace Microsoft.Azure.Commands.Profile.Test
                 cmdlet.ExecuteCmdlet();
                 cmdlet.InvokeEndProcessing();
                 Assert.Equal(ContextSaveMode.CurrentUser, AzureSession.Instance.ARMContextSaveMode);
-                Assert.Equal(typeof(ProtectedFileTokenCache), AzureSession.Instance.TokenCache.GetType());
                 Assert.Equal(typeof(ProtectedProfileProvider), AzureRmProfileProvider.Instance.GetType());
+            }
+            catch (PlatformNotSupportedException)
+            {
+                // swallow exception when test env (Linux server) doesn't support token cache persistence
             }
             finally
             {
@@ -119,7 +125,8 @@ namespace Microsoft.Azure.Commands.Profile.Test
                 cmdlet.ExecuteCmdlet();
                 cmdlet.InvokeEndProcessing();
                 Assert.Equal(ContextSaveMode.Process, AzureSession.Instance.ARMContextSaveMode);
-                Assert.Equal(typeof(AuthenticationStoreTokenCache), AzureSession.Instance.TokenCache.GetType());
+                Assert.True(AzureSession.Instance.TryGetComponent(PowerShellTokenCacheProvider.PowerShellTokenCacheProviderKey, out PowerShellTokenCacheProvider factory));
+                Assert.Equal(typeof(InMemoryTokenCacheProvider), factory.GetType());
                 Assert.Equal(typeof(ResourceManagerProfileProvider), AzureRmProfileProvider.Instance.GetType());
             }
             finally
@@ -142,7 +149,8 @@ namespace Microsoft.Azure.Commands.Profile.Test
                 cmdlet.ExecuteCmdlet();
                 cmdlet.InvokeEndProcessing();
                 Assert.Equal(ContextSaveMode.Process, AzureSession.Instance.ARMContextSaveMode);
-                Assert.Equal(typeof(AuthenticationStoreTokenCache), AzureSession.Instance.TokenCache.GetType());
+                Assert.True(AzureSession.Instance.TryGetComponent(PowerShellTokenCacheProvider.PowerShellTokenCacheProviderKey, out PowerShellTokenCacheProvider factory));
+                Assert.Equal(typeof(InMemoryTokenCacheProvider), factory.GetType());
                 Assert.Equal(typeof(ResourceManagerProfileProvider), AzureRmProfileProvider.Instance.GetType());
             }
             finally