+++ src/KeyVault/KeyVault/Commands/NewAzureKeyVault.cs
@@ -16,11 +16,9 @@ using Microsoft.Azure.Commands.KeyVault.Models;
 using Microsoft.Azure.Commands.KeyVault.Properties;
 using Microsoft.Azure.Commands.ResourceManager.Common.ArgumentCompleters;
 using Microsoft.Azure.Management.KeyVault.Models;
-using Microsoft.WindowsAzure.Commands.Common.CustomAttributes;
 using Microsoft.WindowsAzure.Commands.Utilities.Common;
 using System;
 using System.Collections;
-using System.Linq;
 using System.Management.Automation;
 
 namespace Microsoft.Azure.Commands.KeyVault
@@ -28,9 +26,6 @@ namespace Microsoft.Azure.Commands.KeyVault
     /// <summary>
     /// Create a new key vault.
     /// </summary>
-    [GenericBreakingChange("The ability to create new key vaults with soft delete disabled will be deprecated by December 2020. " +
-        "All key vaults will be required to have soft delete enabled. Please see the following documentation for additional guidance. " +
-        "https://docs.microsoft.com/azure/key-vault/general/soft-delete-change")]
     [Cmdlet("New", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "KeyVault", SupportsShouldProcess = true)]
     [OutputType(typeof(PSKeyVault))]
     public class NewAzureKeyVault : KeyVaultManagementCmdletBase
@@ -86,13 +81,6 @@ namespace Microsoft.Azure.Commands.KeyVault
             ValueFromPipelineByPropertyName = true,
             HelpMessage = "If specified, enables secrets to be retrieved from this key vault by Azure Disk Encryption.")]
         public SwitchParameter EnabledForDiskEncryption { get; set; }
-
-        public const String DisableSoftDeleteChangeDesc = "DisableSoftDelete will be deprecated without being replaced.";
-
-        [CmdletParameterBreakingChange("DisableSoftDelete", "3.0.0", ChangeDescription = DisableSoftDeleteChangeDesc)]
-        [Parameter(Mandatory = false,
-            HelpMessage = "If specified, 'soft delete' functionality is disabled for this key vault.")]
-        public SwitchParameter DisableSoftDelete { get; set; }
 
         [Parameter(Mandatory = false,
             HelpMessage = "If specified, protection against immediate deletion is enabled for this vault; requires soft delete to be enabled as well. Enabling 'purge protection' on a key vault is an irreversible action. Once enabled, it cannot be changed or removed.")]
@@ -170,20 +158,16 @@ namespace Microsoft.Azure.Commands.KeyVault
                     EnabledForDeployment = this.EnabledForDeployment.IsPresent,
                     EnabledForTemplateDeployment = EnabledForTemplateDeployment.IsPresent,
                     EnabledForDiskEncryption = EnabledForDiskEncryption.IsPresent,
-                    EnableSoftDelete = !DisableSoftDelete.IsPresent,
+                    EnableSoftDelete = null,
                     EnablePurgeProtection = EnablePurgeProtection.IsPresent ? true : (bool?)null, // false is not accepted
                     EnableRbacAuthorization = EnableRbacAuthorization.IsPresent,
-
                     /*
-                     * If soft delete is enabled, but retention days is not specified, use the default value,
-                     * else use the vault user provides,
-                     * else use null
+                     * If retention days is not specified, use the default value,
+                     * else use the vault user provides
                      */
-                    SoftDeleteRetentionInDays = DisableSoftDelete.IsPresent
-                        ? null as int?
-                        : (this.IsParameterBound(c => c.SoftDeleteRetentionInDays)
+                    SoftDeleteRetentionInDays = this.IsParameterBound(c => c.SoftDeleteRetentionInDays)
                             ? SoftDeleteRetentionInDays
-                            : Constants.DefaultSoftDeleteRetentionDays),
+                            : Constants.DefaultSoftDeleteRetentionDays,
                     SkuFamilyName = DefaultSkuFamily,
                     SkuName = this.Sku,
                     TenantId = GetTenantId(),