+++ src/Storage/Storage/DatalakeGen2/Cmdlet/GetAzDataLakeGen2Item.cs
@@ -33,7 +33,7 @@ namespace Microsoft.WindowsAzure.Commands.Storage.Blob.Cmdlet
 
         [Parameter(ValueFromPipeline = true, Mandatory = false, HelpMessage =
                 "The path in the specified FileSystem that should be retrieved. Can be a file or directory " +
-                "In the format 'directory/file.txt' or 'directory1/directory2/'. Not specify this parameter to get the root directory of the Filesystem.")]
+                "In the format 'directory/file.txt' or 'directory1/directory2/'. Skip set this parameter to get the root directory of the Filesystem.")]
         [ValidateNotNullOrEmpty]
         public string Path { get; set; }