+++ tools/Az.Tools.Installer/exports/Install-AzModule.ps1
@@ -236,8 +236,15 @@ function Install-AzModule {
         }
 
         if ($RemoveAzureRm -and ($PSCmdlet.ShouldProcess('Remove AzureRm modules', 'AzureRm modules', 'Remove'))) {
-            Uninstall-Module -Name 'AzureRm*' -AllVersion -ErrorAction SilentlyContinue
-            Uninstall-Module -Name 'Azure.*' -AllVersion -ErrorAction SilentlyContinue
+            try {
+                $azureModuleNames = (Get-InstalledModule -Name Azure* -ErrorAction Stop).Name | Where-Object {$_ -match "Azure(\.[a-zA-Z0-9]+)?$" -or $_ -match "AzureRM(\.[a-zA-Z0-9]+)?$"}
+                foreach($azureModuleName in $azureModuleNames) {
+                    Uninstall-Module -Name $azureModuleName -AllVersion -ErrorAction SilentlyContinue
+                }
+            }
+            catch {
+                Write-Warning $_
+            }
         }
 
         #install Az.Accounts first