+++ src/RecoveryServices/RecoveryServices/ChangeLog.md
@@ -18,6 +18,10 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Fixing Workload Restore for contributor permissions.
+* Added new parameter sets and validations for Restore-AzRecoveryServicesBackupItem cmdlet.
+
+## Version 2.12.2
 * Added container name validation for workload backup.
 
 ## Version 2.12.1